Changed how the lengthControl element is found
Before, we were using a jQury find() to get the length control. This
happened not only when the controls were appended to the document - but
also each time the activity was resized. I figured this could have
contributed to poor performance on mobile.

Now, jQuery find() only runs once, and the lengthControl element is
stored in the model where it can be retrieved faster.